client: split out instruction builders to another crate (#526)
