+++ libraries/rust/margin/Cargo.toml
@@ -35,6 +35,7 @@ anchor-client = { git = "https://github.com/jet-lab/anchor", branch = "master" }
 
 jet-simulation = { git = "https://github.com/jet-lab/jet-simulation", branch = "master" }
 jet-program-common = { path = "../program-common" }
+jet-instructions = { path = "../instructions" }
 jet-test-service = { path = "../../../programs/test-service", features = ["no-entrypoint"] }
 jet-fixed-term = { path = "../../../programs/fixed-term", features = ["no-entrypoint", "cli"] }
 jet-airspace = { path = "../../../programs/airspace", features = ["no-entrypoint"] }
