+++ libraries/rust/margin/src/fixed_term/event_consumer.rs
@@ -69,9 +69,10 @@ impl EventConsumer {
                         market: structure,
                         queue: Vec::new(),
                         users: HashMap::new(),
-                        builder: FixedTermIxBuilder::from(structure)
-                            .with_payer(&self.rpc.payer().pubkey())
-                            .with_crank(&self.rpc.payer().pubkey()),
+                        builder: FixedTermIxBuilder::new_from_state(
+                            self.rpc.payer().pubkey(),
+                            structure,
+                        ),
                     })),
                 );
             } else {
@@ -332,7 +333,7 @@ impl MarketState {
 
             let consume_ix = [
                 ComputeBudgetInstruction::set_compute_unit_limit(800_000),
-                self.builder.consume_events(&seed, &consume_params).unwrap(),
+                self.builder.consume_events(&seed, &consume_params),
             ];
             let next_tx = Transaction::new_signed_with_payer(
                 &consume_ix,
