+++ libraries/rust/margin/src/test_service.rs
@@ -30,6 +30,7 @@ use crate::{
     cat,
     fixed_term::FixedTermIxBuilder,
     ix_builder::{
+        fixed_term::OrderBookAddresses,
         get_metadata_address,
         test_service::{
             self, derive_pyth_price, derive_pyth_product, derive_ticket_mint, derive_token_mint,
@@ -313,6 +314,12 @@ fn create_airspace_token_fixed_term_markets_tx(
         let key_bids = Keypair::new();
         let key_asks = Keypair::new();
 
+        let orderbook = OrderBookAddresses {
+            bids: key_bids.pubkey(),
+            asks: key_asks.pubkey(),
+            event_queue: key_eq.pubkey(),
+        };
+
         let len_eq = event_queue_len(EVENT_QUEUE_CAPACITY);
         let len_orders = orderbook_slab_len(ORDERBOOK_CAPACITY);
 
@@ -326,6 +333,7 @@ fn create_airspace_token_fixed_term_markets_tx(
             market_seed,
         ));
         let fixed_term_ix = FixedTermIxBuilder::new_from_seed(
+            config.authority,
             &admin.airspace,
             &mint,
             market_seed,
@@ -333,8 +341,8 @@ fn create_airspace_token_fixed_term_markets_tx(
             derive_pyth_price(&mint),
             derive_pyth_price(&ticket_mint),
             None,
-        )
-        .with_crank(&config.authority);
+            orderbook,
+        );
 
         txs.push(
             test_service::token_register(
@@ -392,22 +400,14 @@ fn create_airspace_token_fixed_term_markets_tx(
                     bm_config.lend_tenor,
                     bm_config.origination_fee,
                 ),
-                fixed_term_ix
-                    .initialize_orderbook(
-                        config.authority,
-                        key_eq.pubkey(),
-                        key_bids.pubkey(),
-                        key_asks.pubkey(),
-                        bm_config.min_order_size,
-                    )
-                    .unwrap(),
+                fixed_term_ix.initialize_orderbook(config.authority, bm_config.min_order_size),
             ],
             signers: vec![key_eq, key_bids, key_asks],
         });
 
         // Submit separately as it is large and causes tx to fail
         txs.push(TransactionBuilder {
-            instructions: vec![fixed_term_ix.authorize_crank(config.authority).unwrap()],
+            instructions: vec![fixed_term_ix.authorize_crank()],
             signers: vec![],
         });
 
@@ -415,7 +415,7 @@ fn create_airspace_token_fixed_term_markets_tx(
             txs.last_mut()
                 .unwrap()
                 .instructions
-                .push(fixed_term_ix.pause_order_matching().unwrap());
+                .push(fixed_term_ix.pause_order_matching());
         }
 
         txs.push(admin.register_fixed_term_market(
